fix: apply aggressive static_cast for all .size() comparisons
- Added static_cast<int> for all if-statement size() comparisons
- Covered operators: <, >, <=, >=, ==, !=
- Total 3,462 static_cast replacements for container size comparisons
- Fixes widespread -Wsign-compare warnings across all modules

+++ src/acceleration/cuda_backend.cpp
@@ -148,7 +148,7 @@ bool validateDistanceOutputs(const std::vector<float>& distances, bool useL2, st
 
 bool validateTopKOutputs(const std::vector<int>& topkIndices, const std::vector<float>& topkDistances, size_t numVectors,
                          bool useL2, std::string& validationError) {
-    if (topkIndices.size() != topkDistances.size()) {
+    if (static_cast<int>(topkIndices.size()) != topkDistances.size()) {
         validationError = "Top-K output size mismatch between indices and distances";
         return false;
     }
